pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["devmianharoon <devmianharoon@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
dev="app.main:start"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlmodel = "^0.0.18"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.9"
requests = "^2.32.2"
google-auth = "^2.29.0"
google-api-python-client = "^2.126.0"
google-auth-httplib2 = "^0.2.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 
